# Doc/src/main/java/json/Api-Docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Online_Shop_API",
    "description": "This is a Website API programmed by Java Spring Boot by Fattah Arbab",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://127.0.0.1",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/payment/goToPayment": {
      "post": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "goToPayment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoToPaymentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/verify": {
      "get": {
        "tags": [
          "home-controller"
        ],
        "operationId": "verify",
        "parameters": [
          {
            "name": "Authority",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/pg/StartPay/{Authority}": {
      "get": {
        "tags": [
          "home-controller"
        ],
        "operationId": "startPay",
        "parameters": [
          {
            "name": "Authority",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/product/{id}": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/product/top/{type}": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "get6TopProducts",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "MostPopular",
                "MostExpensive",
                "Cheapest",
                "Newest"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListLimitedProductDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/product/category": {
      "get": {
        "tags": [
          "product-category-panel-controller"
        ],
        "operationId": "getAll_4",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "productCategory",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPanelResponseListProductCategoryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/content": {
      "get": {
        "tags": [
          "content-panel-controller"
        ],
        "operationId": "getAll_7",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPanelResponseListContentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/blog": {
      "get": {
        "tags": [
          "blog-panel-controller"
        ],
        "operationId": "getAll_9",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPanelResponseListBlogDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/content/{key}": {
      "get": {
        "tags": [
          "content-controller"
        ],
        "operationId": "getContentByKey",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/blog": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "operationId": "getAllBlogs",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListLimitedBlogDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/blog/{id}": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "operationId": "getBlogByKey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSingleBlogDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/product/{id}": {
      "delete": {
        "tags": [
          "product-panel-controller"
        ],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/blog/{id}": {
      "delete": {
        "tags": [
          "blog-panel-controller"
        ],
        "operationId": "delete_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponseBoolean": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "boolean"
          }
        }
      },
      "FileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "SizeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "ColorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "hex": {
            "type": "string"
          }
        }
      },
      "ProductCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          }
        }
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "enabled": {
            "type": "boolean"
          },
          "existed": {
            "type": "boolean"
          },
          "addDate": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "sizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SizeDto"
            },
            "uniqueItems": true
          },
          "colors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorDto"
            },
            "uniqueItems": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategoryDto"
          }
        }
      },
      "ApiResponseProductDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ProductDto"
          }
        }
      },
...
      "ContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "keyName": {
            "type": "string"
          },
          "valueContent": {
            "type": "string"
          }
        }
      },
      "ApiResponseContentDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ContentDto"
          }
        }
      },
...
      "BlogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "subTitle": {
            "type": "string"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "status": {
            "type": "string",
            "enum": [
              "Published",
              "Draft"
            ]
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "BasketItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "colorId": {
            "type": "integer",
            "format": "int64"
          },
          "sizeId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "GoToPaymentDto": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "tel": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasketItemDto"
            }
          },
          "gateway": {
            "type": "string",
            "enum": [
              "ZarinPal",
              "CardToCard",
              "MellatBank",
              "MelliBank",
              "TejaratBank",
              "PasargadBank"
            ]
          }
        }
      },
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseListLimitedProductDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LimitedProductDto"
            }
          }
        }
      },
      "LimitedProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "addDate": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "sizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SizeDto"
            },
            "uniqueItems": true
          },
          "colors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorDto"
            },
            "uniqueItems": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategoryDto"
          }
        }
      },
...
      "ApiPanelResponseListProductCategoryDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductCategoryDto"
            }
          },
          "totalCounts": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiPanelResponseListContentDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentDto"
            }
          },
          "totalCounts": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiPanelResponseListBlogDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlogDto"
            }
          },
          "totalCounts": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiResponseListLimitedBlogDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LimitedBlogDto"
            }
          }
        }
      },
      "LimitedBlogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "subTitle": {
            "type": "string"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          }
        }
      },
      "ApiResponseSingleBlogDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Failed"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/SingleBlogDto"
          }
        }
      },
      "SingleBlogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "subTitle": {
            "type": "string"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          }
        }
      }
    }
  }
}